+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 * Provides `PlatformInterface` with common mechanism for enforcing that a platform interface
   is not implemented with `implements`.
 * Provides test only `MockPlatformInterface` to enable using Mockito to mock platform interfaces.
+
+## 0.0.2 - Test case added - 21 February 2025
+
+* Added `fake_webview_platform` implementation for enforcing that a webview platform interface test 
+* `webview_platform_interface_test` and a `example` project.